Last Changes
Changed the shape of the array to something more shape like (more of a
challenge instead of just 3 dots)
Commented the code
Edited the code to work with the new shape
Added some tones
…
I'm going to need someone else to upload my game to their Meggy. My
computer is extremely out of whack and is almost unable to upload every
time.